refactor: Validator 클래스로 모든 유효성 검사 로직 분리
